+++ HISTORY.md
@@ -2,6 +2,9 @@
 
 ### v0.0.1-alpha
 
-> 此版本发布于 2222-02-22
+> 此版本发布于 2022-10-28
 
-* xxx
+* 支持 DES/3DES/AES 对称加密算法
+* 支持 ECB/CBC/OFB/CFB/CTR 等分组模式
+* 支持 PKCS5/PKCS7/ZERO/NO 等填充方式
+* 支持 HEX/BASE64 等编解码算法